+++ Sources/OchaCore/Watcher.swift
@@ -47,20 +47,3 @@ public class Watcher {
         FSEventStreamStart(stream)
     }
 }
-
-
-public final class FileEvent : NSObject {
-    public var id: UInt64
-    public var flag: UInt32
-    public var path: Pathable
-
-    public init(
-        id: UInt64,
-        flag: UInt32,
-        path: Pathable
-        ) {
-        self.id = id
-        self.flag = flag
-        self.path = path
-    }
-}
